requirements.txt
azure-functions==1.8.0
Pillow==9.0.0
pytesseract==0.3.8
